fix error about check app install statue, sometimes can't get the install result

+++ AUICrawler/controller/appController.py
@@ -14,34 +14,36 @@ sys.setdefaultencoding('utf-8')
 
 
 def install_app(device, apk_path):
-    Saver.save_crawler_log(device.logPath, 'Step : install app : ' + apk_path)
-    command = 'adb -s ' + device.id + " install -r " + apk_path
-    result = os.popen(command).readlines()
-    result = result[-1]
-    if 'Success' in result:
-        del device, apk_path, command, result
-        return True
-    else:
-        print result
-        del device, apk_path, command, result
-        return False
+    try:
+        if os.path.exists(apk_path):
+            Saver.save_crawler_log(device.logPath, 'Step : install app : ' + apk_path)
+            command = 'adb -s ' + device.id + " install -r " + apk_path
+            os.system(command)
+            del device, apk_path, command
+    except:
+        del device, apk_path
+        Saver.save_crawler_log(device.logPath, 'install app catch exception')
 
 
 def uninstall_app(device, package_name):
-    Saver.save_crawler_log(device.logPath, 'Step : uninstall app : ' + package_name)
-    command = 'adb -s ' + device.id + " uninstall " + package_name
-    os.system(command)
-    del device, package_name, command
+    try:
+        Saver.save_crawler_log(device.logPath, 'Step : uninstall app : ' + package_name)
+        command = 'adb -s ' + device.id + " uninstall " + package_name
+        os.system(command)
+        del device, package_name, command
+    except:
+        del device, package_name
+        Saver.save_crawler_log(device.logPath, 'uninstall app catch exception')
 
 
 def app_is_installed(device, package_name):
     Saver.save_crawler_log(device.logPath, "Step : check app is installed or not")
-    command = 'adb -s ' + device.id + " shell pm list"
+    command = 'adb -s ' + device.id + " shell pm list package"
     result = os.popen(command)
     lines = result.readlines()
     for line in lines:
-        if ("package:"+package_name) == line[:-2]:
-            Saver.save_crawler_log(device.logPath, "app is installed")
+        if package_name in line and (package_name + '.') not in line:
+            print "app is installed"
             del command, lines, device, line, package_name
             return True
         del line
